fix: good form tests setup
